Refactor Terraform codebase: - Introduced new `writability.go` file to handle writable property checks and related logic. - Removed legacy `generator.go` file, consolidating functionality into the new structure. - Deleted `naming.go` file, utilizing the internal naming package directly.
