+++ internal/openapi/writability_test.go
