+++ internal/terraform/generate_locals.go
