+++ internal/terraform/generate_outputs.go
@@ -1,0 +1,24 @@
+package terraform
+
+import (
+	"github.com/hashicorp/hcl/v2/hclwrite"
+	"github.com/matt-FFFFFF/tfmodmake/internal/hclgen"
+	"github.com/zclconf/go-cty/cty"
+)
+
+func generateOutputs() error {
+	file := hclwrite.NewEmptyFile()
+	body := file.Body()
+
+	resourceID := body.AppendNewBlock("output", []string{"resource_id"})
+	resourceIDBody := resourceID.Body()
+	resourceIDBody.SetAttributeValue("description", cty.StringVal("The ID of the created resource."))
+	resourceIDBody.SetAttributeRaw("value", hclgen.TokensForTraversal("azapi_resource", "this", "id"))
+
+	name := body.AppendNewBlock("output", []string{"name"})
+	nameBody := name.Body()
+	nameBody.SetAttributeValue("description", cty.StringVal("The name of the created resource."))
+	nameBody.SetAttributeRaw("value", hclgen.TokensForTraversal("azapi_resource", "this", "name"))
+
+	return hclgen.WriteFile("outputs.tf", file)
+}
